+++ Dyalect/Compiler/UnitComposition.cs
@@ -13,10 +13,16 @@ namespace Dyalect.Compiler
         {
             Units = units;
             Types = StandardType.All.Clone();
+            Members = new FastList<string>();
+            MembersMap = new Dictionary<string, int>();
         }
 
         public List<Unit> Units { get; }
 
         public FastList<DyTypeInfo> Types { get; }
-    }    
+
+        internal FastList<string> Members { get; }
+
+        internal Dictionary<string, int> MembersMap { get; }
+    }
 }