+++ Docs/Changes.md
@@ -1,3 +1,9 @@
+# 0.3.4
+  * A bug fixed - incorrect type info were generated for the values of type `Array`.
+  * Runtime type system is refactored.
+  * Optimizations in members functions execution.
+  * Now module import directives are always compiled before the rest of the code.
+
 # 0.3.3
   * Logic, responsible for the implementation of function calls, is largely rewritten and optimized.
   * Internal call stack structure is changed for better effeciency.